refactor: move the original directory 'cmd/app/*' to 'pkg/cmd/gtctl'
