+++ pkg/cmd/gtctl/cluster/delete/delete.go
