+++ pkg/cmd/gtctl/cluster/scale/scale.go
