+++ pkg/cmd/gtctl/version/version.go
